requirements.txt
pyTelegramBotAPI
google-generativeai
requests
pytz
